moving splashPage
to pages
