+++ src/pages/SplashPage/SplashScreenElements.js
