+++ src/pages/SplashPage/index.js
